+++ packages/mask/src/plugins/Wallet/Dashboard/index.tsx
@@ -1,7 +1,7 @@
 import type { Plugin } from '@masknet/plugin-infra'
 import { base } from '@masknet/plugin-wallet'
 import { ExtensionSite, getSiteType, NetworkPluginID } from '@masknet/shared-base'
-import { SelectNftContractDialog } from '../SNSAdaptor/SelectNftContractDialog.js'
+import { SelectNftContractDialog } from '../SNSAdaptor/SelectNftContractDialog/index.js'
 import { SelectProviderDialog } from '../SNSAdaptor/SelectProviderDialog/index.js'
 import { WalletStatusDialog } from '../SNSAdaptor/WalletStatusDialog/index.js'
 import { ConnectWalletDialog } from '../SNSAdaptor/ConnectWalletDialog/index.js'
