+++ packages/mask/src/plugins/Wallet/SNSAdaptor/SelectNftContractDialog/index.tsx
@@ -9,9 +9,9 @@ import { OpenInNew as OpenInNewIcon } from '@mui/icons-material'
 import type { NonFungibleCollection } from '@masknet/web3-shared-base'
 import { useChainContext, useNonFungibleCollections } from '@masknet/web3-hooks-base'
 import { WalletMessages } from '@masknet/plugin-wallet'
-import { useI18N } from '../../../utils/index.js'
-import { SearchInput } from '../../../extension/options-page/DashboardComponents/SearchInput.js'
 import { FuseNonFungibleCollection } from '@masknet/web3-providers'
+import { useI18N } from '../../../../utils/index.js'
+import { SearchInput } from '../../../../extension/options-page/DashboardComponents/SearchInput.js'
 
 const useStyles = makeStyles()((theme) => ({
     search: {
