+++ packages/mask/src/plugins/Wallet/SNSAdaptor/index.tsx
@@ -2,7 +2,7 @@ import { getEnumAsArray } from '@masknet/kit'
 import { NetworkPluginID } from '@masknet/shared-base'
 import type { Plugin } from '@masknet/plugin-infra'
 import { base } from '@masknet/plugin-wallet'
-import { SelectNftContractDialog } from './SelectNftContractDialog.js'
+import { SelectNftContractDialog } from './SelectNftContractDialog/index.js'
 import { SelectProviderDialog } from './SelectProviderDialog/index.js'
 import { WalletStatusDialog } from './WalletStatusDialog/index.js'
 import { ConnectWalletDialog } from './ConnectWalletDialog/index.js'
