+++ packages/web3-providers/src/types/Telemetry.ts
